requirements.txt
waitress
flask
flask_sqlalchemy
flask_login
werkzeug
pillow
opencv-python